Move metric trigger tests to proper test file
- Create test_automations_schema.py for client-side schema validation tests
- Move metric trigger tests from server compatibility file
- Remove unnecessary deferred imports and comments
- These tests validate client-side deserialization, not server compatibility
